requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
distro==1.9.0
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2024.3.1
future==1.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.0
idna==3.6
importlib_metadata==7.1.0
Jinja2==3.1.3
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai==1.14.2
openai-whisper==20231117
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pvrecorder==1.2.2
PyAudio==0.2.14
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pynput==1.7.6
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==306
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.1
sympy==1.12
tdqm==0.0.1
tiktoken==0.6.0
tk==0.1.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.39.1
typing_extensions==4.10.0
urllib3==2.2.1
Wave==0.0.2
whisper_mic==1.4.0
zipp==3.18.1
